style(components/select-time-range): move to `/common` and fix `onChange` `value` type
