+++ src/components/common/index.ts
@@ -1,3 +1,4 @@
 export * from './open-in-spotify-button'
 export * from './audio-bars'
 export * from './infinite-loading-button'
+export * from './select-time-range'
